hud: move webview to pkg
This is so that we can reference protobuf files that are going to live
here in the future.
