+++ pkg/hud/webview/analytics_nudge.go
