+++ pkg/hud/webview/testdata/snapshot.json
