+++ pkg/hud/webview/testdata/view.json
